Call the tracker druks and fold its board into the app's own modules.
The tracker choice is now `druks`, not `issues`, so the stored value, the
work item source, the prompts, and the frontend all read the same word.

The board has no package of its own. Its statuses joined `enums.py`, its
schemas `schemas.py`, its routes `routes.py`, its pages are the app's
`pages.py`, and `DruksTracker` sits in `ticketing/` next to Linear and Jira.
Its tables are `tickets` and `ticket_comments`, beside the app's other
unprefixed tables.

The ticket routes answer like the rest of the agent surface. A blank title
or comment is a wire constraint the platform enforces, an unknown ticket,
repo, or account raises `TicketNotFound`, `RepoNotFound`, or
`OwnerNotFound`, and each route declares what it can refuse.
`CreateProjectRequest` held one field, so `create_project` takes an
embedded body instead.

The build prompts no longer single out this tracker. They told the agent
which tool to call and that the reference is not a GitHub issue number
because `issues` read as GitHub Issues; the word is gone, so one sentence
covers every tracker again. `TRACKER_LABELS` went with it: the prompts name
the tracker by the source value, which is also the name of the MCP server
holding its tools.

One migration replaces the stack's seven: it adds the project prefix and
sequence, backfills a prefix per project, creates both ticket tables, and
adds `personal_access_tokens.allowed_tools`.

Build.dispatch asks `SubjectStatus.is_active` instead of listing run
states, and the board tests cover the same behaviour in 518 lines instead
of 747.
