+++ backend/druks/durable/schemas.py
@@ -13,7 +13,7 @@ from pydantic import (
 
 from druks.schemas import Schema
 
-from .enums import AgentCallStatus, RunState
+from .enums import ACTIVE_STATES, AgentCallStatus, RunState
 
 if TYPE_CHECKING:
     from .models import Run
@@ -173,6 +173,10 @@ class SubjectStatus(Schema):
     # has only the driving run, which is what everything above already comes from.
     triggered_at: datetime | None = None
     account_username: str | None = None
+
+    @property
+    def is_active(self) -> bool:
+        return self.state in ACTIVE_STATES
 
     @property
     def is_parked(self) -> bool:
